Added Property Class
Changed properties to be stored in a class
separated pre-existing classes into new files
